fixed python3 issue with tf.cast

+++ code/utils/weightnorm.py
@@ -84,7 +84,8 @@ class AdamWithWeightnorm(Adam):
         if self.initial_decay > 0:
             lr *= (1. / (1. + self.decay * self.iterations))
 
-        t = self.iterations + 1
+        t = K.cast(self.iterations + 1, 'float32')
+
         lr_t = lr * K.sqrt(1. - K.pow(self.beta_2, t)) / (1. - K.pow(self.beta_1, t))
 
         shapes = [K.get_variable_shape(p) for p in params]
